Initial and un-run version of force-cut figure S2. IS UNTESTED.
In more detail, although this is an untested version of the file, I do
think that it shouldn't take too much debugging. One of the primary
concerns is that it is dependent on the new (and also untested)
quiver_cut method of the FigureConstructor.

This part is unimportant: The reason that this is untested is that it
was written on an airplane with a tiny compy and as such a test would be
prohibitively long and battery draining.

 --
 user: Dave Williams <cdave@dante.u.washington.edu>
 added better_practices/FigureS2.py