Great speed up of crossbridge.py using std lib's math.
The functions in the math module of the standard library appear to be vastly faster than those of numpy. Perhaps because they are just c functions, perhaps because they aren't intended to deal with vector values. In any event, switching to them dropped a sample execution from 255 seconds to 44 seconds. Also, using a different optimizer was nice.